Added information regarding parameter initialization for Simple API function calls
Previously, the Simple API documentation did not explicitly note the requirement that all parameters to Simple API function calls must be initialized and allocated prior to each call.

This requirement is now noted in the Simple API introduction, and relevant function documentation has been updated with a note recommending review of the introduction for information about parameter initialization.