Use minideb instead of the default GitLab image
I wasn't able to find clear documentation on the default image
(https://gitlab.com/gitlab-org/gitlab-runner/-/issues/1577 suggests
`ruby:2.1`?) but I think it's safe to assume it will be more than 60 MB.

This is still not the absolute minimum size, the `alpine:latest` image
is only 6 MB, but that requires upstream support from cargo-deadlinks
that doesn't yet exist.

This brings the build time down from 2 minutes and 7 seconds to 1 minutes and 47 seconds.