[Pipeline] Specify image for pages job
We moved our pipelines locally; our set-up uses Alpine linux as a
default; not Ubuntu or Debian. This was apparently implicitly assumed by
the script; and it's not a correct assumption. A job should specify
which image it could use. Here, we use `apt-get` as part of the
before-script, and therefore assume Ubuntu/Debian.