[Pipeline] Fix re-clone error for ../YDBOcto
We implemented local pipelines that implement caching. That means that a
repos could have been already downloaded before. Error as follows:

```
$ git clone --depth 1 https://gitlab.com/YottaDB/DBMS/YDBOcto ../YDBOcto
fatal: destination path '../YDBOcto' already exists and is not an empty
directory.
```

This change applies some logic to fetch the changes if the repo already
exists; otherwise, it will download it again as previously.