[Pipeline] Octo documentation was not the latest
Because of the mechanics of the internal pipeline on our machine, if the
docker volumes that contain the artifact are still around, we may not
get the latest code in the working tree as the code is only doing a git
fetch not a git pull. This commit fixes this omission. This was an error
in the commit 637a0ccfd62eb95eb1928d89206984caf9de905c.