[#348] Revert parts of c38c9c commit
This commit says:

```
- Fixed logo and contents bullet list issue by fixing sphinx_rtd_theme to v0.5.2
```

It seems we don't need to fix the sphinx_rtd_theme, as that had the side
effect of breaking search.

Search broke because the htmlpatch.csh file couldn't perform this step:

```
 89         # Step 4
 90         set filelist = `ls -1 _build/html/search.html`
 91         set from = '<script src="_static/js/theme.js"></script>'
 92         set to1 = '<script type="text/javascript" src="_static/language_data.js"></script>'
 93         set to = "$from${to1}"
 94         perl -p -i -e "s|$from|$to|g" $filelist
```

This is because the specific version chosen (0.5.2) for
sphinx_rtd_theme seemed to be pretty old, and had an older syntax for
for the the from line, and thus the search replacement didn't work.

Also, installing Sphinx from PIP installs Sphinx v6, which does not show
the logo. This is a problem that is still not addressed. So the part of
c38c9c is also reverted.