[#351] Make htmlpatch.csh more robust in case it fails to make the changes requested
Commit 07c00947019235bdf876fcba72b14ab5c5baf47a (almost 2 years ago),
which added `tcsh -fe` made this possible, but the sed changes never
fail if there is nothing to be replaced; which caused problems as some
of the changes never got deployed to the pipeline when Sphinx versions
changed and the code we touched didn't exist anymore in Sphinx.

The fix is to add `grep -q` statements. grep fails if there are no
matches, thus causing the make to stop.

Also added `set verbose` so it would be obvious in the pipeline if the
whole script fails to run; but in any case, the exit status would be 1
in that case due to the grep failure.