[#277] Created example on how to pass TCP sockets between processes
Previously, there was no documentation providing an example of how to pass TCP sockets between processes in M. This commit adds such an example program to demonstrate this operation for users.

Specifically, new `passTCPSocket.m` and `passTCPSocket2.m` demo programs were added to the M Programmer's Guide in the *Socket Device Examples* section in `ioproc.rst`, along with a brief introduction and explanation of each program's operation. Links to download the programs as M files were also added for users to easily acquire and run the programs for themselves.

Also, added `build/` and `deadlinks` to `.gitignore` to declutter `git status` output in YDBDoc.

Finally, a note on the convention of using `,!` to flush sockets was added to `ioproc.rst` docs. Also, the preceding paragraph to this addition was revised to wrap code snippets in backticks so that they display as code on the web.