[#356] Added Octo and GUI to Acculturation Guide
Previously, the Acculturation Guide omitted instructions on installing Octo or YDBGUI. This commit adds a new section for each to the `Using YottaDB` section of `acculturation.rst`.

The Acculturation Guide `Quickstart` section and `Virtual Machine Configuration` sections were revised to include Octo and GUI installation options and port forwarding for Octo and YDBGUI, respectively.

Both of the new Octo and YDBGUI usage sections point to the newly revised `Quickstart` and `Virtual Machine Configuration` sections for instructions on installing the plugins in the YottaDB virtual machine.

The new Octo section includes setup and usage instructions, as well as examples.

The YDBGUI section demonstrates both how to run YDBGUI from the terminal, as well as how to access it from the web. A sample screenshot is also included for reference.

Additionally, various usages of `pkg-config` to set `ydb_dist` were replaced with a hardcoded value to reflect the default YottaDB install path used by `ydbinstall.sh`.