[#348][#353] Fixed documentation to work with latest Sphinx version and removed `htmlpatch.csh`
Previously, the documentation build process relied on `htmlpatch.csh` to manually override various Sphinx theme styles as well as insert scripts and html into the build. However, this approach resulted in breakage when using newer versions of Sphinx, which invalidated the assumptions made in `htmlpatch.csh`.

To resolve this breakage and provide a forward-compatible way to override theme styles, etc., this commit makes the following changes:
1. Specify style overrides in the pre-existing `custom.css` file of each documentation section
2. Added the yottadb pixel to `custom.css` as a `background-image` on `html` tags, instead of adding it via `htmlpatch.csh`

Note that the script insertions done by `htmlpatch.csh` are not, in fact necessary, as these scripts and the functionality they support are not maintained or modified by us, but included automatically by Sphinx itself. So, these are simply removed with the deletion of `htmlpatch.csh`. This omission was confirmed to be acceptable through manual testing of the search functionality of the documentation pages, which these scripts are required to support.