[#355] Added Applications Manual with YDBApron tutorial
YDBDoc#355 requests the creation of an applications manual with a guidance on developing applications using YottaDB across a variety of domains. This commit creates this manual and adds a tutorial for developing sample application in the web domain.

Specifically, a new `ApplicationsManual` directory was added to the documentation, and the `Makefile` and `buildall.sh` files were updated to include the new directory in the documentation build process.

Moreover, a new `ydbapron.rst` file containing a tutorial on building the YDBApron sample application was added to the `ApplicationsManual` directory. This tutorial provides an overview of the YDBApron application and its data model, as well as a file-by-file breakdown of its implementation. Screenshots are also included, along with a diagram illustrating the application data model and YottaDB tree structure.

The tutorial also includes an exhaustive set of code samples that concretely demonstrate how the YottaDB API can be used in a web application.

Additionally, a new `ydbpython.rst` file was added with YDBPython introductory tutorials.

Miscellaneous:
+ Added 'Mode of Address' section to StyleGuide
+ Revised needlessly complicated `font-family` attribute across `custom.css` files for simplicity