[#355] Added 'YottaDB and a tale of two Pis' to the Applications Manual
YDBDoc#355 requests the creation of an applications manual with a guidance on developing applications using YottaDB across a variety of domains. This commit creates this manual and adds a tutorial for developing a sample application in the IoT domain.

Specifically, a new `demopi.rst` file was added, containing instructions on a simple IoT setup using YottaDB on two Raspberry Pis. This tutorial includes a number of embedded images, which were also added to the repository, along with a `demopi.py` file containing the full code for the application.