[#364] Added Copy Button to Sphinx Documentation
This commit adds a copy button to code blocks in the documentation for user convenience. This is done by adding the `sphinx_copybutton` extension to the extension list in the `conf.py` file in each documentation section.

So that the copy button is now included in pipeline builds, the `python3-sphinx-copybutton` package is now included in `.gitlab-ci.yml`.