[#361] Reduce duplication in documentation source files
Previously, YDBDoc contained a lot of duplicated files. This commit removes the duplication by maintaining a single copy of the relevant files in a new `shared` folder.

To ensure these files are included at build time, `Makefile` was modified to link these files to the relevant manual directories before building the documentation with Sphinx.

Much of `conf.py` is duplicated, however there is no straightforward way to maintain only the unique portions of this file, since the unique portions are embedded within the duplicate portions. At this point, it was deemed not worth attempting to separate these things out programmatically.

Nonetheless, the various `conf.py` files were revised to remove superficial discrepancies, such that now the only differences between these files are significant.

Note that `LICENSE.rst` was not consistent across manuals, so the most recent version was chosen to be retained and shared across all manuals.

Finally, `.gitignore` was removed as none of the files should be
ignored.