[#358] Miscellaneous small fixes and enhancements to user documentation post r1.38 through r2.00
Change documentation of UTF-8 support in the Language Extensions chapter of the M Programmers Guide to remove discussion about $ydb_dist needing to point to the utf8 subdirectory in UTF-8 mode. Since M routines in YottaDB distributions are always in libyottadbutil.so, it is $ydb_routines that needs to point to the correct shared library; $ydb_dist is irrelevant. Also, add some explanatory discussion.

Additionally:

- Make documentation of messages DLLCHSETM and SLLCHSETUTF8 more helpful.
- Refer to LC_ALL and LC_CTYPE as locale settings, matching Linux man page language
- Clarify / correct relationship between locale command and locale settings LC_ALL and LC_CTYPE. The previous language was potentially misleading at best, and borderline wrong at worst.
- Correct typos.
- Responses to review comments.

Responses to review comments