[#390] Update copyright year in files used to build YottaDB documentation
Commit !725 missed copyrights in files used to build the documentation. Although those files themselves have not changed in years, since the copyright text in them is displayed in the published documentation, they need show the copyright year of the published documentation rather than the dates that they themselves were changed. This commit updates the copyright information in those files.