Do not depend on the git state of other repositories
YDBDoc depends on other git repositories for several things:
- YDBLua, for docs/lua-yottadb-ydbdocs.rst.
- YDBOcto, for the `doc/` folder.
- YDB, for checking with `ci/error-sync.py` whether the error codes are up to date.

Previously, it looked at "raw" git state in the other repositories, and would try to pull branches from the remote. This disrupted work in those repos, and did not work at all if HEAD was not on a branch.

Instead, use a local checkout in `git-checkouts/` which is managed by the script and does not have any of these problems.