Compare to the merge request target when verifying copyrights.
Previously, we used `commit_verify.sh` with its default target branch,
which is `master`. But YDBDoc does not use master for development, it
uses tagged version names (currently, r2.04). And as a result, we would
get errors in the first week of January for files that had been modified
in the last week of December. Use CI_MERGE_REQUEST_TARGET_BRANCH_NAME
instead, which should fix the error, at the downside of only running the
commit-verify script when you open an MR, not immediately when you push.
I think this is an ok tradeoff, I don't see people using long-lasting
feature branches for YDBDoc.