Fix `needs-clone.sh` when there is no local checkout of a git repo
There are two ways `ci/needs-clone.sh` creates a git worktree in `./git-checkouts`. The first is when the git repo is already cloned locally to (e.g.) `../lua-yottadb`, in which case it uses a git worktree. The second is when it is not yet cloned, in which case it clones the remote itself.

In that second case, it failed to set `upstream/HEAD` to a commit, because git does not do so automatically when adding a new remote.
Do so manually.