Disable error check until I can get it working reliably
This check compares the error reference to YDB master. But YDBDoc master refers to the latest
published release of YDB, not the latest development commit. I tried disabling it in CI when making an MR against master,
but it is also causing issues when building locally. Disable the check for now.