Do not use shallow fetches when reusing an existing git repository
In https://gitlab.com/YottaDB/DB/YDBDoc/-/merge_requests/800, I added a `needs-clone.sh` script which runs on `make test` to ensure that all the necessary git repositories are checked out. That script uses an existing local git repo if it exists, and if so runs `git fetch --depth 1` to get the tip of upstream.

I expected that to fetch the latest commit, and leave the local history untouched. But it turns out that even if a full history already exists locally, `fetch --depth 1` will convert it to a shallow history. As a result, pushing to the remote becomes impossible, resulting in errors like this one:
```
$ git push origin +master
To git+ssh://gitlab.com/ksbhaskar/YDB.git
 ! [remote rejected]   master -> master (shallow update not allowed)
```

Avoid this by only using shallow clones if the repository is fully managed by `needs-clone`, not if it is an existing checkout.

Note that this prevents the problem from reoccuring in the future, but does not fix any existing history. Converting back from shallow to full history is a manual operation and can be done as follows:
```
git fetch --unshallow upstream
```