Consistently use the correct filename in `check_missing_errors`
`check_missing_errors` previously used `args.rst` and now uses the `filename` parameter. `args.rst` comes from the CLI parsing code on line 491, and in practice always defaults to `MessageRecovery/errors.rst` on line 47. But this error path is taken not just when checking errors.rst, but also when checking errormsgref.rst. And so sometimes it would print this incorrect message:
```
Checking MessageRecovery/errormsgref.rst is in sync with git-checkouts/YDB/sr_port/*.msg

ERROR: STACKRANGE was documented in git-checkouts/YDB/sr_port/ydberrors.msg:82 but not in MessageRecovery/errors.rst
```
After this change, `filename` is determined by the caller of `check_missing_errors`, and you can see on line 457 that it passes in `args.err_msg_ref`, not `args.rst`.