[#417][#444] Improved the error message when a git branch is not based off of r2.04 or master
Previously, when a branch was based off of the `r2.04` branch and a change was made to the `r2.04` branch but not incorporated into the derivative branch, an error message like the following would be emitted:

```sh
$ ci/error-check.sh
ERROR: HEAD should be based off [r2.04] or [master] branch.
ERROR: This script will not work correctly until the situation is fixed.
```

With this commit, an error message like the following is now emitted, which explains the underlying issue and suggests a fix for it:

```
$ ci/error-check.sh
ERROR: HEAD should be based off [r2.04] or [master] branch.
ERROR: This script will not work correctly until the situation is fixed.
NOTE: HEAD is based off an old version [f52a7952c0c3760f76130f6b31f9bd05355f2d11] of [r2.04]. Consider rebasing: 'git rebase upstream/r2.04'
```