[#442] Use a single unified sphinx project
- fix broken LICENSE links
- remove duplicate files from .gitignore
- after unifying, sphinx gave a warning:
```
/home/jyn/work/YDBDoc/AdminOpsGuide/dbrepl.rst:279: WARNING: duplicate citation a, other instance in /home/jyn/work/YDBDoc/ProgGuide/ioproc.rst
/home/jyn/work/YDBDoc/AdminOpsGuide/dbrepl.rst:281: WARNING: duplicate citation b, other instance in /home/jyn/work/YDBDoc/ProgGuide/ioproc.rst
```

It turns out that `[a]` is a "citation", not a "footnote", and hence can
be referred to across documents, and so must be globally unique.
Switch to `[#a]`, which only needs to be unique per-page.

This changes the rendered note from a `b` superscript to a `2`
superscript; hopefully that's ok.

- Fix a broken link in isv.rst
- Disambiguate duplicate links
- Add a chapter heading to each subproject

This shows a more useful title in the sidebar on the left,
and also prevents "Indices and tables" from being shown.

- fix caching and octo; see inline comments for more detail
- switch from hand-coded index.html to index.rst

this matches the style of the rest of the docs and is nicer to maintain.
we need something like it anyway; sphinx would allow a separate toctree
but that seems like extra work.