[#417][#453] Restored navigation bar in Octo documentation
Previously, the navigation bar for all YottaDB documentation was missing from the Octo documentation page. This was because `buildall.sh` built Octo separately then `rsync`d it into the separate Sphinx build done for the rest of the documentation.

This commit fixes this issue by building Octo along with the rest of the documentation, instead of separately. This is done by adding a new `octo` rule to the `Makefile` to copy the Octo documentation source files from YDBOcto *before* building YDBDoc and including them in the YDBDoc build with all the other documentation guides.

This makes the Octo-specific logic in `buildall.sh` unnecessary, so that is now removed.

Also:
+ `.gitignore` was updated to include directories only used during the Octo portion of the YDBDoc build
+ `make clean` was similarly updated to clean up new files and directories created during the YDBDoc build process for the Octo documentation