[#457] Removed "1" literal from YottaDB documentation navigation
Previously a "1" literal appeared at the top of the YottaDB documentation navigation below the logo but above the search bar. This "1" is output as the current documentation version as specified by the `version` variable in `conf.py`.

However, this value is incorrect, in that the YottaDB documentation version mirrors the upstream YottaDB version. At the same time, updates to the YottaDB documentation are published outside of the YottaDB release cycle, making this version number irrelevant for the online documentation.

Moreover, the current YDBDoc build process doesn't generate any independent documentation files like PDFs or EPUB files.

In view of the above and the fact that the "1" literal in the documentation navigation is unsightly, this commit removes it by setting `version = ""` in `conf.py`.