[#417][#469] Fix documentation for ydb_cip[_t]
ydb_cip
-------
The documentation for `ydb_cip` neglected to mention that the `ydb_string_t` passed into the function should have a null terminated string as its address
and that the length field is ignored. The documentation now clarifies that despite being a struct with a char* and a length value, the string should be null terminated.

ydb_cip_t
---------
Similarly, the documentation for `ydb_cip_t` was also incorrect.
The documentation mentioned a non-existent type `ci_rtn_name` as the 3rd argument and while it did mention that it needed a null terminated string,
the way the struct that contains that pointer is built was unclear.
It now clearly states that the ydb_string_t in the ci_name_descriptor should contain a null terminated string
in the address and that the ydb_string_t's length is ignored.